[fix] PR 멘토링 fetch 부분 제외 모두 반영
