+++ public/index.html
@@ -1,17 +1,20 @@
 <!doctype html>
 <html lang="ko">
-    <head>
-        <meta charset="utf-8" />
-        <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <meta property="og:title" content="Linkbrary" />
-        <meta property="og:description" content="세상의 모든 정보를 쉽게 저장하고 관리해보세요" />
-        <meta property="og:image" content="./img/library.jpg" />
-        <meta property="og:url" content="giho-linkbrary.netlify.app" />
-        <meta name="twitter:card" content="summary" />
-        <title>Linkbrary</title>
-    </head>
-    <body>
-        <div id="root"></div>
-    </body>
+  <head>
+    <meta charset="utf-8" />
+    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta property="og:title" content="Linkbrary" />
+    <meta property="og:description" content="세상의 모든 정보를 쉽게 저장하고 관리해보세요" />
+    <meta property="og:image" content="./img/library.jpg" />
+    <meta property="og:url" content="giho-linkbrary.netlify.app" />
+    <meta name="twitter:card" content="summary" />
+    <title>Linkbrary</title>
+  </head>
+  <body>
+    <div id="root"></div>
+    <noscript>
+      <div>자바스크립트 실행하지 못했습니다...</div>
+    </noscript>
+  </body>
 </html>
