+++ src/App.css
@@ -1,13 +1,13 @@
 :root {
-    --privary: #6D6AFE;
-    --red: #FF5B56;
-    --black: #111322;
-    --white: #FFFFFF;
-    --gray1: #3E3E43;
-    --gray2: #9FA6B2;
-    --gray3: #CCD5E3;
-    --gray4: #E7EFFB;
-    --gray5: #F0F6FF;
+  --privary: #6d6afe;
+  --red: #ff5b56;
+  --black: #111322;
+  --white: #ffffff;
+  --gray1: #3e3e43;
+  --gray2: #9fa6b2;
+  --gray3: #ccd5e3;
+  --gray4: #e7effb;
+  --gray5: #f0f6ff;
 }
 
 @import url('/src/assets/pretendard.css');
