+++ src/components/CardList/Card/Card.css
@@ -1,95 +1,94 @@
 .link-list {
-    display: grid;
-    grid-template-rows: repeat(3, 340px);
-    grid-template-columns: repeat(3, 340px);
-    gap: 20px;
+  display: grid;
+  grid-template-rows: repeat(3, 340px);
+  grid-template-columns: repeat(3, 340px);
+  gap: 20px;
 }
 
 .link {
-    width: 340px;
-    height: 334px;
-    box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
-    border-radius: 15PX;
-    overflow: hidden;
+  width: 340px;
+  height: 334px;
+  box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
+  border-radius: 15px;
+  overflow: hidden;
 }
 
 .link__img {
-    width: 100%;
-    height: 205px;
-    overflow: hidden;
+  width: 100%;
+  height: 205px;
+  overflow: hidden;
 }
 
 .link__img img {
-    width: 100%;
-    height: 100%;
-    object-fit: contain;
-    transition: all 0.1s;
+  width: 100%;
+  height: 100%;
+  object-fit: contain;
+  transition: all 0.1s;
 }
 
 .link:hover .link__img img {
-    transform: scale(1.3);
+  transform: scale(1.3);
 }
 
 .link:hover .link__content {
-    background: var(--gray5);
+  background: var(--gray5);
 }
 
 .link__content {
-    box-sizing: border-box;
-    width: 100%;
-    height: 135px;
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    padding: 15px 20px;
-    background: #FFF;
+  box-sizing: border-box;
+  width: 100%;
+  height: 135px;
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  padding: 15px 20px;
+  background: #fff;
 }
 
 .link__content--time-passed {
-    color: #666;
-    font-family: Pretendard;
-    font-size: 13px;
+  color: #666;
+  font-family: Pretendard;
+  font-size: 13px;
 }
 
 .link__content--description {
-    overflow: hidden;
-    color: #000;
-    font-family: Pretendard;
-    font-size: 16px;
-    text-overflow: ellipsis;
-    display: -webkit-box;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
-    -webkit-line-clamp: 2;
+  overflow: hidden;
+  color: #000;
+  font-family: Pretendard;
+  font-size: 16px;
+  text-overflow: ellipsis;
+  display: -webkit-box;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
+  -webkit-line-clamp: 2;
 }
 
 .link__content--createdAt {
-    overflow: hidden;
-    color: #333;
-    text-overflow: ellipsis;
-    font-family: Pretendard;
-    font-size: 14px;
+  overflow: hidden;
+  color: #333;
+  text-overflow: ellipsis;
+  font-family: Pretendard;
+  font-size: 14px;
 }
 
 /* tablet */
-@media (max-width:1199px) {
-    .link-list {
-        display: grid;
-        grid-template-columns: repeat(2, 340px);
-        gap: 24px;
-    }
+@media (max-width: 1123px) {
+  .link-list {
+    display: grid;
+    grid-template-columns: repeat(2, 340px);
+    gap: 24px;
+  }
 }
 
 /* smartPhone */
-@media (max-width:767px) {
-    .link-list {
-        display: grid;
-        grid-template-columns: 100%;
-        gap: 24px;
-    }
+@media (max-width: 767px) {
+  .link-list {
+    display: grid;
+    grid-template-columns: 100%;
+    gap: 24px;
+  }
 
-    .link {
-        width: 100%;
-        height: auto;
-    }
+  .link {
+    width: 100%;
+  }
 }
