+++ src/components/CardList/CardList.css
@@ -2,9 +2,9 @@
 @import url('./Card/Card.css');
 
 .links {
-    padding: 40px 32px;
-    margin-bottom: 60px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
+  padding: 40px 32px;
+  margin-bottom: 60px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
