+++ src/components/CardList/Search/Search.css
@@ -1,36 +1,35 @@
 .cards__search-bar {
-    box-sizing: border-box;
-    width: 1060px;
-    padding: 15px 16px;
-    border-radius: 10px;
-    background: #F5F5F5;
-    display: flex;
-    gap: 10px;
-    align-items: center;
-    margin-bottom: 40px;
+  box-sizing: border-box;
+  width: 1060px;
+  padding: 15px 16px;
+  border-radius: 10px;
+  background: #f5f5f5;
+  display: flex;
+  gap: 10px;
+  align-items: center;
+  margin-bottom: 40px;
 }
 
 .cards__search-bar img {
-    width: 16px;
-    height: 16px;
+  width: 16px;
+  height: 16px;
 }
 
-.cards__search-bar>input {
-    width: 100%;
-    background-color: #F5F5F5;
+.cards__search-bar > input {
+  width: 100%;
+  background-color: #f5f5f5;
 }
-
 
 /* tablet */
-@media (max-width:1199px) {
-    .cards__search-bar {
-        width: 704px;
-    }
+@media (max-width: 1123px) {
+  .cards__search-bar {
+    width: 704px;
+  }
 }
 
 /* smartPhone */
-@media (max-width:767px) {
-    .cards__search-bar {
-        width: 100%;
-    }
+@media (max-width: 767px) {
+  .cards__search-bar {
+    width: 100%;
+  }
 }
