+++ src/components/Folder/Folder.css
@@ -1,32 +1,33 @@
 .folder {
-    background: var(--gray5);
-    padding: 113px 0 60px;
+  width: 100vw;
+  background: var(--gray5);
+  padding: 113px 0 60px;
 }
 
 .folder--container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
 
 .folder--container img {
-    width: 60px;
-    height: 60px;
-    border-radius: 50%;
-    background: var(--primary-purple-70, #8F00FF);
-    margin-bottom: 12px;
+  width: 60px;
+  height: 60px;
+  border-radius: 50%;
+  background: var(--primary-purple-70, #8f00ff);
+  margin-bottom: 12px;
 }
 
 .folder--container span {
-    color: var(--text-color-light-mode, #000);
-    font-family: Pretendard;
-    font-size: 16px;
-    margin-bottom: 20px;
+  color: var(--text-color-light-mode, #000);
+  font-family: Pretendard;
+  font-size: 16px;
+  margin-bottom: 20px;
 }
 
 .favorites {
-    color: #000;
-    font-family: Pretendard;
-    font-size: 40px;
-    font-weight: 600;
+  color: #000;
+  font-family: Pretendard;
+  font-size: 40px;
+  font-weight: 600;
 }
