+++ src/components/Footer/Footer.css
@@ -1,58 +1,58 @@
 footer {
-    box-sizing: border-box;
-    width: 100vw;
-    padding: 32px 104px 64px;
-    background: #111322;
+  box-sizing: border-box;
+  width: 100vw;
+  padding: 32px 104px 64px;
+  background: #111322;
 }
 
 .footer-container {
-    max-width: 1712px;
-    display: grid;
-    grid-template-areas: 'codeit privacy-policy-and-faq link';
-    justify-content: space-between;
-    margin: 0 auto;
+  max-width: 1712px;
+  display: grid;
+  grid-template-areas: 'codeit privacy-policy-and-faq link';
+  justify-content: space-between;
+  margin: 0 auto;
 }
 
 .footer-font {
-    color: #676767;
-    font-family: Abel;
-    font-size: 16px;
+  color: #676767;
+  font-family: Abel;
+  font-size: 16px;
 }
 
 .footer-codeit {
-    grid-area: codeit;
+  grid-area: codeit;
 }
 
 .footer-privacy-policy-and-faq {
-    grid-area: privacy-policy-and-faq;
-    display: flex;
-    gap: 30px;
+  grid-area: privacy-policy-and-faq;
+  display: flex;
+  gap: 30px;
 }
 
 .footer-link {
-    grid-area: link;
-    display: flex;
-    align-items: center;
+  grid-area: link;
+  display: flex;
+  align-items: center;
 }
 
 .footer-link a:not(:first-of-type) {
-    margin-left: 12px;
+  margin-left: 12px;
 }
 
 /* mobile */
-@media (max-width:767px) {
-    footer {
-        padding: 32px;
-    }
+@media (max-width: 767px) {
+  footer {
+    padding: 32px;
+  }
 
-    .footer-container {
-        grid-template-areas:
-            'privacy-policy-and-faq link'
-            'codeit empty';
-        align-items: start;
-    }
+  .footer-container {
+    grid-template-areas:
+      'privacy-policy-and-faq link'
+      'codeit empty';
+    align-items: start;
+  }
 
-    .footer-privacy-policy-and-faq {
-        margin-bottom: 60px;
-    }
+  .footer-privacy-policy-and-faq {
+    margin-bottom: 60px;
+  }
 }
