+++ src/components/Header/Header.css
@@ -1,81 +1,81 @@
 .header {
-    box-sizing: border-box;
-    width: 100%;
-    padding: 32px 200px;
-    background: var(--gray5);
-    position: fixed;
-    z-index: 9999;
+  box-sizing: border-box;
+  width: 100%;
+  padding: 32px 200px;
+  background: var(--gray5);
+  position: fixed;
+  z-index: 9999;
 }
 
 .header--container {
-    max-width: 1920px;
-    display: flex;
-    justify-content: space-between;
-    margin: 0 auto;
+  max-width: 1920px;
+  display: flex;
+  justify-content: space-between;
+  margin: 0 auto;
 }
 
 .header__img img {
-    width: 100%;
-    height: 100%;
+  width: 100%;
+  height: 100%;
 }
 
 .header__profile {
-    display: flex;
-    align-items: center;
+  display: flex;
+  align-items: center;
 }
 
 .header__profile img {
-    width: 28px;
-    height: 28px;
-    border-radius: 50%;
-    margin-right: 6px;
+  width: 28px;
+  height: 28px;
+  border-radius: 50%;
+  margin-right: 6px;
 }
 
 .header__profile div {
-    color: var(--Linkbrary-gray100, #373740);
-    font-family: Pretendard;
-    font-size: 14px;
+  color: var(--Linkbrary-gray100, #373740);
+  font-family: Pretendard;
+  font-size: 14px;
 }
 
 .header__btn-login {
-    width: 128px;
-    border-radius: 8px;
-    padding: 16px 20px;
-    color: var(--Grey-Light, #F5F5F5);
-    font-family: Pretendard;
-    font-size: 18px;
-    font-weight: 600;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 10px;
-    background: var(--gra-purpleblue-to-skyblue, linear-gradient(91deg, #6D6AFE 0.12%, #6AE3FE 101.84%));
+  width: 128px;
+  border-radius: 8px;
+  padding: 16px 20px;
+  color: var(--Grey-Light, #f5f5f5);
+  font-family: Pretendard;
+  font-size: 18px;
+  font-weight: 600;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
+  background: var(--gra-purpleblue-to-skyblue, linear-gradient(91deg, #6d6afe 0.12%, #6ae3fe 101.84%));
 }
 
 /* tablet */
-@media (max-width:1199px) and (min-width: 768px) {
-    .header {
-        padding: 20px 32px;
-    }
+@media (max-width: 1123px) and (min-width: 768px) {
+  .header {
+    padding: 20px 32px;
+  }
 
-    .header--container {
-        width: 800px;
-    }
+  .header--container {
+    width: 800px;
+  }
 }
 
-@media (max-width:864px) {
-    .header--container {
-        width: auto;
-    }
+@media (max-width: 864px) {
+  .header--container {
+    width: auto;
+  }
 }
 
 /* smartPhone */
-@media (max-width:767px) {
-    .header {
-        padding: 13px 32px;
-    }
+@media (max-width: 767px) {
+  .header {
+    padding: 13px 32px;
+  }
 
-    .header--container {
-        width: 100%;
-    }
+  .header--container {
+    width: 100%;
+  }
 }
