+++ src/reset.css
@@ -1,26 +1,26 @@
 * {
-    margin: 0;
-    padding: 0;
-    font: inherit;
-    color: inherit;
+  margin: 0;
+  padding: 0;
+  font: inherit;
+  color: inherit;
 }
 
 button {
-    background: none;
-    border: 0;
-    cursor: pointer;
+  background: none;
+  border: 0;
+  cursor: pointer;
 }
 
 a {
-    text-decoration: none;
-    cursor: pointer;
+  text-decoration: none;
+  cursor: pointer;
 }
 
 input {
-    outline: none;
-    border: none;
+  outline: none;
+  border: none;
 }
 
 li {
-    list-style: none;
+  list-style: none;
 }
